refactor(chat): extract acp status bar state
